requirements.txt
streamlit==1.30.0
python-dotenv==1.0.0
groq==0.4.1
langchain==0.1.0
langchain-groq==0.1.0
langchain-core==0.1.0
pydantic==1.10.8
zipfile36==0.1.3
pathlib==1.0.1
numpy==1.26.2
pandas==2.1.4